Moved Feedback component from New Furniture to Homepage component.
