+++ src/components/common/ClientFeedback/ClientFeedback.module.scss
@@ -14,6 +14,7 @@
 
   .icon {
     font-size: 20px;
+    margin-top: 30px;
   }
 
   .description {
