+++ src/components/views/Homepage/Homepage.js
@@ -1,15 +1,15 @@
 import React from 'react';
 // import PropTypes from 'prop-types';
-
 import styles from './Homepage.module.scss';
-
 import FeatureBoxes from '../../features/FeatureBoxes/FeatureBoxes';
 import NewFurniture from '../../features/NewFurniture/NewFurnitureContainer';
+import Feedback from '../../features/Feedback/FeedbackContainer';
 
 const Homepage = () => (
   <div className={styles.root}>
     <FeatureBoxes />
     <NewFurniture />
+    <Feedback />
   </div>
 );
